git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6436 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHFemMesh.cpp
@@ -354,7 +354,7 @@ void PHFemMesh::SetDesc(const void* p) {
 			}
 */
 			//外側の法線だけ加算
-			if(vertices[vid].faces[fid] < nSurfaceFace){
+			if(vertices[vid].faces[fid] < (int)nSurfaceFace){
 				vertices[vid].normal += faces[vertices[vid].faces[fid]].normal;		// このコードに代わって、上記vectorコードと以下の加算コードに置き換え
 			}
 				/*			if(faces[vertices[vid].faces[fid]].normal == 0){
